OpenCL version uploaded.

+++ SpGEMM_opencl/SpGEMM_EM_kernels.cl
@@ -13,7 +13,7 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 
 #define TUPLE_QUEUE 6
-typedef double   vT;
+typedef float   vT;
 
 inline
 void binarysearch(__local int   *s_key,
