+++ SpGEMM_opencl/SpGEMM_ESC_2heap_kernels.cl
@@ -12,7 +12,7 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 
 #define TUPLE_QUEUE 6
-typedef double   vT;
+typedef float   vT;
 
 inline
 void siftDown(__local int   *s_key,
